refactor: move _testing.go into _test.go file
